add support for having Banchan-specific currency names
Fixes: https://github.com/BanchanArt/banchan/issues/639

Also did some refactoring of currency-related stuff into a Banchan.Payments.Currency module
